+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## [2.1.0] - 2019-06-13
 - Added addExtension method to TwigViewRenderer.
 - Added setDebug method to TwigViewRenderer.
 
@@ -11,5 +11,6 @@
 ## 1.0.0 - 2018-02-03
 - First stable revision.
 
+[2.1.0]: https://github.com/themichaelhall/bluemvc-twig/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/themichaelhall/bluemvc-twig/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/themichaelhall/bluemvc-twig/compare/v1.0.0...v1.1.0